refactor: 122.5a -- give the pointer-motion decision its own module
122.5 made the pane-resolution chain pure and testable but left it in
`app_impl.rs`, which GREW to 5,784 lines as a result: the extraction
added ~450 lines of tests to the god file it was supposed to relieve.

One coherent concept accounted for ~900 of those lines -- the
out-of-frame pointer-motion repaint decision -- so it now has a module
whose path names it, per `freminal-module-cohesion`. This is the case
the `freminal-extend-or-extract` skill (122.0) was written for; the
alternative was leaving it in the file it was already too big for.

Moved to `freminal/src/gui/pointer_motion.rs`: `PointerMotionPaneSignals`,
`PaneSnapshotInputs`, `PaneResolution`, `pane_hover_region_risk`,
`pane_hover_region_terms`, `pointer_in_gutter_strip`,
`animation_in_flight_composed`, `pointer_motion_needs_repaint_decision`,
`resolve_pane_under_pointer`, and their 35 tests.

  app_impl.rs   5784 -> 4878   (-906)
  pointer_motion.rs        945 (new)

A pure move: no function body changed except visibility keywords, no
test assertion changed, nothing renamed, and the test count is
identical at 1166 before and after.

Visibility is the narrowest that compiles. Three of the six functions
(`pane_hover_region_risk`, `pane_hover_region_terms`,
`pointer_in_gutter_strip`) stayed fully private -- nothing outside the
module calls them directly -- and the rest are `pub(super)`. The module
is deliberately NOT `pub`, even though that would have unblocked the
benchmark 122.5 had to skip: that repeats the trade the 122.14
amendment rejected, and cohesion is the justification here, not
benchmarking.

`App::pointer_motion_needs_repaint` and `App::is_chrome_interactive_at`
stay in `app_impl.rs` as trait-impl methods and call into the new
module.

Also fixes two doc references this move made stale: `window.rs`'s
`PointerMotionConditionFlags` comparison pointed at `app_impl.rs` for
`PointerMotionPaneSignals`, and the 122.14 benchmark's module doc
explained the excluded predicates in terms of `mod app_impl`'s
visibility and cited a line range in that file for their tests. Both
now name `gui::pointer_motion`, and the benchmark's exclusion rationale
covers the extracted chain too.

Verification: cargo test --all; cargo test --all --features
frame-profiling; cargo clippy --all-targets --all-features -D warnings;
cargo machete; cargo fmt --check; cargo bench --no-run --all.
